backend/requirements.txt
fastapi
uvicorn
python-multipart
python-magic
PyYAML
docker
pandas